feat: reduce redundancy of grouping symbols (#11)
In some cases groupings are used to indicate that a certain part of
expression should be interpreted as a single term. For example in
fractions: `(a + b) / (c + d)`, or when using scripts `x^(n-1)` and so
on.

In such cases, the grouping symbols don't have to be rendered. For
example $\frac{a + b}{c + d} \equiv \frac{(a + b)}{(c + d)}$. The same
holds for the scripts: $x^{n-1} \equiv x^{(n-1)}$.

This PR solves that problem. Also, some structures are simplified by
using `SimpleExpr::Interm` instead of `SimpleExpr::Grouping` with
ignored grouping symbols. This makes snapshots somewhat simpler.

Additionally, block rendering in binary is now supported with `-b,
--block` flag.